+++ packages/generate-material/src/constants.ts
@@ -1,0 +1,7 @@
+export const ENV_MAP = {
+  web: 'isWeb',
+  weex: 'isWeex',
+  miniapp: 'isMiniApp',
+  'wechat-miniprogram': 'isWechatMiniProgram',
+  kraken: 'isKraken',
+};
